+++ src/managed.rs
@@ -6,6 +6,7 @@ pub fn is_compatible_extension_version(version: &str) -> bool {
         || version == "0.35.0"
         || version == "0.36.0"
         || version == "0.37.0"
+        || version == "0.38.0"
         || version == "1.0.0"
     {
         return true;
@@ -165,6 +166,7 @@ mod tests {
         assert!(is_compatible_extension_version("0.35.0"));
         assert!(is_compatible_extension_version("0.36.0"));
         assert!(is_compatible_extension_version("0.37.0"));
+        assert!(is_compatible_extension_version("0.38.0"));
         assert!(is_compatible_extension_version("1.0.0-rc.1"));
         assert!(is_compatible_extension_version("1.0.0-rc.2"));
         assert!(is_compatible_extension_version("1.0.0-rc.42"));